Refactor prepayments
